requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
python-subunit>=0.18.0
fixtures>=3.0.0
six>=1.10.0
testtools>=2.2.0
